+++ backend/builder/tests/conftest.py
@@ -1,0 +1,9 @@
+import pytest
+
+import orchestrator
+import orchestrator.mock as orc_mock
+
+
+@pytest.fixture
+def function() -> orchestrator.Function:
+    return orc_mock.FunctionFactory()