+++ backend/substrapp/compute_tasks/errors.py
@@ -71,10 +71,10 @@ class ExecutionError(_ComputeTaskError, CeleryNoRetryError):
 
     def __init__(self, logs: BinaryIO, *args, **kwargs):
         self.logs = logs
-        super().__init__(*args, **kwargs)
+        super().__init__(logs, *args, **kwargs)
 
 
-def get_error_type(exc: Exception) -> failure_report_pb2.ErrorType:
+def get_error_type(exc: Exception) -> failure_report_pb2.ErrorType.ValueType:
     """From a given exception, return an error type safe to store and to advertise to the user.
 
     Args: