+++ backend/substrapp/tasks/tasks_save_image.py
@@ -1,6 +1,9 @@
+from __future__ import annotations
+
 import os
 import pathlib
 from tempfile import TemporaryDirectory
+from typing import Any
 
 import structlog
 from django.conf import settings
@@ -11,8 +14,10 @@ from backend.celery import app
 from image_transfer import make_payload
 from substrapp.compute_tasks import utils
 from substrapp.docker_registry import USER_IMAGE_REPOSITORY
+from substrapp.models import FailedAssetKind
 from substrapp.models import FunctionImage
-from substrapp.tasks.tasks_compute_task import ComputeTask
+from substrapp.orchestrator import get_orchestrator_client
+from substrapp.tasks.task import FailableTask
 
 REGISTRY = settings.REGISTRY
 REGISTRY_SCHEME = settings.REGISTRY_SCHEME
@@ -21,22 +26,50 @@ SUBTUPLE_TMP_DIR = settings.SUBTUPLE_TMP_DIR
 logger = structlog.get_logger("worker")
 
 
+class SaveImageTask(FailableTask):
+    autoretry_for = settings.CELERY_TASK_AUTORETRY_FOR
+    max_retries = settings.CELERY_TASK_MAX_RETRIES
+    retry_backoff = settings.CELERY_TASK_RETRY_BACKOFF
+    retry_backoff_max = settings.CELERY_TASK_RETRY_BACKOFF_MAX
+    retry_jitter = settings.CELERY_TASK_RETRY_JITTER
+    acks_late = True
+    reject_on_worker_lost = True
+    ignore_result = False
+
+    asset_type = FailedAssetKind.FAILED_ASSET_FUNCTION
+
+    @property
+    def attempt(self) -> int:
+        return self.request.retries + 1  # type: ignore
+
+    # Returns (function key, channel)
+    def get_task_info(self, args: tuple, kwargs: dict) -> tuple[str, str]:
+        function = orchestrator.Function.parse_raw(kwargs["function_serialized"])
+        channel_name = kwargs["channel_name"]
+        return function.key, channel_name
+
+    # Celery does not provide unpacked arguments, we are doing it in `get_task_info`
+    def on_success(self, retval: dict[str, Any], task_id: str, args: tuple, kwargs: dict[str, Any]) -> None:
+        function_key, channel_name = self.get_task_info(args, kwargs)
+        with get_orchestrator_client(channel_name) as client:
+            client.update_function_status(
+                function_key=function_key, action=orchestrator.function_pb2.FUNCTION_ACTION_READY
+            )
+
+
 @app.task(
     bind=True,
     acks_late=True,
     reject_on_worker_lost=True,
     ignore_result=False,
-    base=ComputeTask,
+    base=SaveImageTask,
 )
 # Ack late and reject on worker lost allows use to
 # see http://docs.celeryproject.org/en/latest/userguide/configuration.html#task-reject-on-worker-lost
 # and https://github.com/celery/celery/issues/5106
-def save_image_task(self: ComputeTask, function_serialized: str, channel_name: str, function_key: str) -> None:
+def save_image_task(task: SaveImageTask, function_serialized: str, channel_name: str) -> tuple[str, str]:
     logger.info("Starting save_image_task")
-    logger.info(
-        f"Parameters: function_serialized {function_serialized}, "
-        f"channel_name {channel_name}, function_key {function_key}"
-    )
+    logger.info(f"Parameters: function_serialized {function_serialized}, " f"channel_name {channel_name}")
     # create serialized image
     function = orchestrator.Function.parse_raw(function_serialized)
     container_image_tag = utils.container_image_tag_from_function(function)
@@ -60,3 +93,5 @@ def save_image_task(self: ComputeTask, function_serialized: str, channel_name: s
             function_id=function.key, file=File(file=storage_path.open(mode="rb"), name="image.zip")
         )
         logger.info("Serialized image saved")
+
+    return function_serialized, channel_name